Port SpiderMonkey to wasm32-unknown-unknown (Portable Baseline Interpreter)
Adds a wasm32-unknown-unknown target configuration for the mozjs_sys build
so it can be linked into a Cloudflare Worker: recognizes the target triple
in config.sub/init.configure, disables native JIT tiers, shared memory,
tests, and shell tooling, and forces the Portable Baseline Interpreter
(the SpiderMonkey execution tier documented for environments where
runtime code generation is unavailable) per the design in servo-wasm's
docs/wasm-javascript.md.

Provides Worker-safe implementations for the platform primitives that
don't exist without an OS: GC memory mapping (gc/Memory.cpp), executable-
memory allocation (jit/ProcessExecutableMemory.cpp), signal handlers
(WasmSignalHandlers.cpp, MmapFaultHandler.*), threading (noop mutex/
condvar for this target, matching the existing WASI noop path), and a
monotonic TimeStamp backend (TimeStamp_worker.cpp) driven by a host-
supplied nanosecond clock import, wired up via a servo_worker_monotonic_now_ns
FFI entry point on the Rust side.

Adjusts the bindgen pipeline (mozjs-sys/build.rs) for this target: exposes
hidden wasm declarations during binding generation, patches a handful of
generated bindings for libclang/wasm-specific representation mismatches,
and links against wasi-sysroot's libc/libc++ purely for their
implementations (not for WASI itself -- see the follow-up commit
eliminating the resulting WASI imports).

+++ mozjs-sys/mozjs/js/src/vm/BytecodeUtil.cpp
@@ -2792,7 +2792,7 @@ static bool GetPCCountJSON(JSContext* cx, const ScriptAndCounts& sac,
 
     json.beginObject();
 
-    json.property("id", offset);
+    json.property("id", static_cast<uint64_t>(offset));
     json.property("line", range.frontLineNumber());
     json.property("name", CodeName(op));
